// CMIS4-FE/src/app/modules/compoents-customer-module/models/thong-tin-khach-hang.model.ts
export class TTinKHang {
    ID_KHANG: String;
    ID_DDO: String;
    MA_KHANG: String;
    TEN_KHANG: String;
    DTHOAI: String;
    MA_HDONG: String;
    MA_KHTT: String;
    WEBSITE: String;
    EMAIL: String;
    MANHOM_KHANG: String;
    LOAI_KHANG: String;
    DIA_CHI_KH: String;
    MA_DDO: String;
    SO_HO: Number;
    LOAI_DDO: Number;
    DIA_CHI_DDO: String;
    KIMUA_CSPK: Number;
    SO_PHA: Number;
    MA_SOGCS: String;
    KVUC_STT: String;
    MA_TRAM: String;
    MA_LO: String;
    MA_TO: String;
    MA_LOAIHD: String;
    MA_CAPDA: String;
    NGAY_HLUC_DDO: String;
    THAO_TACDL: String;
    SO_COT: String;
    SO_HOP: String;
    TEN_TRAM: String;
    TEN_CAPDA: String;
    TENNHOM_KHANG: String;
    TEN_LOAIHD: String;
    TENLOAI_DDO: String;
    NGAY_HHLUC: String;
}